refactor: 프로젝트 상세 페이지 컴포넌트 분리
