+++ src/pages/MainPage.tsx
@@ -1,5 +1,5 @@
 import MainBanner from 'components/main/banner/MainBanner';
-import MainCalendar from 'components/main/calendar/MainCalendar';
+import MainCalendar from 'components/main/Calendar/MainCalendar';
 import WebContainer from 'components/common/WebContainer';
 import MainRecommendation from 'components/main/recommendation/MainRecommendation';
 const MainPage = () => {
